BFF API directory refactor
